+++ Objtree.h
@@ -154,10 +154,10 @@ typedef struct Point3D
 // Plane definitions
 typedef struct Plane
 {
-    float           A;              // Vector of plane equation
-    float           B;              // A(x-x1) + B(y-y1) + C(z-z1) = 0 or Ax1 + By1 + Cz1 + D = 0
-    float           C;              // D component = -(Ax1 + By1 + Cz1) if used
-    float           D;              // ABCD are passed as array to glClipPlane
+    double           A;              // Vector of plane equation
+    double           B;              // A(x-x1) + B(y-y1) + C(z-z1) = 0 or Ax1 + By1 + Cz1 + D = 0
+    double           C;              // D component = -(Ax1 + By1 + Cz1) if used
+    double           D;              // ABCD are passed as array to glClipPlane
     struct Point    refpt;          // Point (x1,y1,z1) that lies on the plane
 } Plane;
 
@@ -165,10 +165,10 @@ typedef struct Plane
 // If only the ABC are used, it is interchangeable with a Plane.
 typedef struct PlaneRef
 {
-    float           A;              // Vector of plane equation
-    float           B;              // A(x-x1) + B(y-y1) + C(z-z1) = 0 or Ax1 + By1 + Cz1 + D = 0
-    float           C;
-    float           D;
+    double           A;              // Vector of plane equation
+    double           B;              // A(x-x1) + B(y-y1) + C(z-z1) = 0 or Ax1 + By1 + Cz1 + D = 0
+    double           C;
+    double           D;
     struct Point*   refpt;         // Pointer to a point that lies on the plane
 } PlaneRef;
